Transformation of numeric values into colors

Transformation of numeric values into colors

Usage

indicators_to_colors(
  v_indicator,
  v_worst,
  v_neutral,
  v_best,
  col_worst = "#E6550D",
  col_neutral = "#EFF3FB",
  col_best = "#08306B",
  col_NA = "#00000000",
  resolution_bad = 0.1,
  resolution_good = 0.1,
  resolution_legend = 1
)

Arguments

v_indicator

Numeric Vector of indicator data

v_worst

Worst possible value

v_neutral

Neutral value

v_best

Best possible value

col_worst

Color for worst value as hexadecimal string of the form "#rrggbb"

col_neutral

Color for worst value as hexadecimal string of the form "#rrggbb"

col_best

Color for worst value as hexadecimal string of the form "#rrggbb"

col_NA

Color for NA values (Transparent by default)

resolution_bad

the resolution between neutral and worst value

resolution_good

The resolution between netral and best value

resolution_legend

The resolution of the legend values

Value

Vector of colors corresponding to the indicator values